Add colortable.c
