+++ Source/smokeview/colortable.c
@@ -1,0 +1,44 @@
+#include "options.h"
+
+#include <stdio.h>
+#include <string.h>
+#include GLUT_H
+#include <math.h>
+
+#include "smokeviewvars.h"
+#include "IOscript.h"
+#include "MALLOCC.h"
+#include "glui_smoke.h"
+#include "glui_bounds.h"
+#include "histogram.h"
+
+
+/* ------------------ GetColorTableIndex ------------------------ */
+
+int GetColorTableIndex(int *color){
+  int i;
+
+  if(colortableinfo==NULL)return -1;
+  for(i=0;i<ncolortableinfo;i++){
+    colortabledata *cti;
+
+    cti = colortableinfo + i;
+    if(color[0]==cti->color[0]&&color[1]==cti->color[1]&&color[2]==cti->color[2])return i;
+  }
+  return -1;
+}
+
+/* ------------------ GetColorTable ------------------------ */
+
+colortabledata *GetColorTable(char *label){
+  int i;
+
+  if(label==NULL||strlen(label)==0)return NULL;
+  for(i=0;i<ncolortableinfo;i++){
+    colortabledata *cti;
+
+    cti = colortableinfo + i;
+    if(strcmp(label,cti->label)==0)return cti;
+  }
+  return NULL;
+}
